Extract query-method inner functions to `query::inner`
